Tudushka/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases[postgresql]
asyncpg
pydantic
psycopg2-binary
python-jose
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
email-validator
alembic
python-multipart
